Moved board edge 0.5 mm closer to P1
The USB Type-C connector was a bit far from the board edge, so I moved
the edge 0.5 mm closer to it.  This allows enclosure walls to be thicker
and whatnot.  I also moved the screw holes back to keep them the same
distance from the board edge, and moved D4 and R10 to keep their text
unobscured.