Thicker high-current traces
All the power traces immediately around the MOSFET are now zones.
Huzzah!  The power traces at the USB connector are slightly wider now
too.  And to top it off, the ground connection to the power output
connector now has a separate zone to remove the thermal relief instead
of wide traces.