Changed long VBUS trace to zone
The long VBUS trace connecting the USB connector to the MOSFET was a bit
marginal in terms of width, and plenty of room was available for it.
Therefore, I decided to change it into a filled zone, which is at least
5 mm wide for most of its length.  The resistance will be much lower,
and all the vias stitching it to the other VBUS zone under the MOSFET
will help carry heat away as well.