4-layer board with smaller components
This is a major change!  Almost all the components have changed to
smaller versions, with 0603 and 0402 passives and a QFN microcontroller.
The board now has four layers, with internal layers used for VBUS and
GND.  The board now measures a mere 25×30 mm!  The SWD header and Boot
switch footprints have been moved to the bottom to help with this
shrinkage.  Two test points were added to the bottom as well to make it
possible to power the board without using the USB connector.  The
mounting holes are now slightly smaller (better sized for M3), and only
17 mm apart instead of the old 18 mm.

To sum up, everything is smaller and that's almost entirely a good
thing, except now it's not really possible to build one by hand.