Use new USB Type-C schematic symbol
The footprints haven't been merged yet, but the symbol has.  Therefore,
I updated the footprint in the pd-buddy library to match the new symbol
(changed shield pin number to S1) and changed the schematic to have the
new connector symbol.  The board looks the same except that a couple
references changed.