Added programming jig PCB
It's a small board, just a bit bigger than a PD Buddy Sink v0.1.
Nothing unexpected, just pogo pins and connectors and two resistors
because it has a USB Type-C receptacle for power input.  The plastic
piece screws on and the whole thing should work just great, making the
SWD header unnecessary on the v0.3 PCBs.  It's even compatible with the
boot switch, and with a 2-pin 0.1 in. male header on the bottom as an
output connector.