Update footprints
The SOT-89 footprint changed its name slightly in the KiCad library,
so the voltage regulator has been updated accordingly.  It has a nice
new 3D model too.  It's not weird and blue anymore.  I like it.

The SOD-123 was also updated to have a 3D model.  The board now only
lacks models for the footprints I made myself.  Heh.