Removed copper from between SW1 pins 2 and 3
In the previous board version, there was a trace (OUT_CTRL) between the
Boot switch's pins 2 and 3.  Since those two pins are the ones that
need to be shorted to enter DFU mode, it's possible that users wanting
to upgrade the device's firmware would want to short them with a blob
of solder rather than buying a switch.  While the soldermask *should*
prevent this solder blob from shorting the OUT_CTRL line high,
soldermask can be accidentally scraped off, so I don't want to rely on
it in this case.

This commit moves the OUT_CTRL trace out from between SW1 pins 2 and 3,
and adds a keepout zone between those two pins.  This should make it
much safer to bridge those pins with a blob of solder.