Changed SW1 and P1 footprints
The components on the bottom, SW1 and P1, were using slightly modified
footprints from the KiCad library (default, but with no paste on any
pads) in v0.3.  Now they're back to the default footprints, but with
local paste clearance set to -100% to make sure there's no paste on any
of their pads.