Changed Q2 and D1
D1 is now a B140, which is a MacroFab house part.  It's actually a bit
more expensive from distributors than the SS14FL, but the lack of a
labor charge makes it quite a bit cheaper for PCBA.  It's also larger,
so in order to fit the new part, I moved it between Q2 and J1.  This is
possible because…

Q2 is now a DMP3017SFG.  This is a cheaper, smaller, and less
over-specified MOSFET which should be fine for the job of controlling
the PD Buddy Sink's output.  It uses a proprietary Diodes Inc. package
whose footprint I contributed to KiCad's library.  Because the new
MOSFET is so much smaller than the old one, I was able to fit it close
to the board edge where R9 used to be.

The whole output circuit has been rearranged, and I like its layout a
lot more now.  Also, the new position of D1 allowed me to move a bunch
of reference designators closer to their footprints, which is nice.