Added test points, replaced SW1 with R11
More test points will be needed for automated testing, so I've added the
first two I know of.

With firmware version 1.2.0+, there's not much reason to install SW1
anymore.  This commit replaces it with R11, an unpopulated resistor
footprint that does the same thing as bridging the two pins of SW1 used
to do, just in case an upgrade make the `boot` command stop working.